requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
peewee==3.17.5
pyTelegramBotAPI==4.18.1
requests==2.32.2
telebot==0.0.5
urllib3==2.2.1
python-dotenv~=1.0.1